Changes to lesson 01
Besides minor changes in the form (rewording) and
presentation/organization of the document, I propose some other changes
that address more important issues. Next I briefly describe some of
those changes:

1. I moved the illustration of the ‘help’ command just after the Search
Documentation bullet. ‘help’ is just another way to access MATLAB
documentation.

2. I found the analogies: variable <-> sticky note, and assignment <->
putting the sticky note on a particular value, quite misfortunate.
Without the aim of entering in a philosophical discussion, I simply
consider that these analogies are wrong and they do not illustrate what
actually happens in reality. Essentially, we don’t label a particular
value with a name (the sticky note), this would imply that you have an
infinite space of values (some of them labeled, some of them not), and
that you label those that you get to use… What we really have is a
finite space of memory and you fill this memory with the value that you
need. In this sense, the classical analogy where the variable is seen
as a container where you can put a particular value (via the
assignment) is the one that we could use if we want to. But I rather
don’t use any analogy, so I removed this part from the document.

3. MATLAB does store all data in the form of arrays. An array,
continuing with the analogies, is a box or container with a shape and a
size (BTW shape and size are not synonyms), so, you cannot have a box
with zero dimensions! A number is stored as the single element of an
array of one column and one row. A vector is stored as an array with
multiple rows or multiple columns, and a Matrix is stored as an array
with multiple rows and multiple columns.

4. We want to give immediate reward to our lectors; to give them, right
away, ideas of how to solve problems that they normally encounter in
their day-to-day work. The title of the lesson is “Analyzing Patient
Data”, and I like the idea of having this problem as the conducting
thread of the discussion. I was surprised, then, when the big part of
the lesson that explain how to access specific data in an array, was
illustrated with a Magic Square matrix and not with the actual matrix
of inflammation data that has been loaded into memory. I modified this
part to work over the ‘patient_data’ array instead. I left, however,
the Magic Square for the last two selections: the ‘every third row’ and
the ‘checkerboard’, because a) I like Magic Squares, b) I think the
included graphs are really useful, and c) I found kind of odd to
justify such selection patterns in the patients’ inflammation data
context.

5. I modified the part where the various statistics are presented. I
believe it is clearer to put first the presentation of the mean
function across the different dimensions, and then the operation over
the entire array. One thing that is important to clarify is that
‘patient_data(:)’ not only flattens the array, but it rearranges the
elements of the matrix in column-major order.  This means that the data
corresponding to the same patient is now scattered by a 40 spaces
stride, and not consecutively as we may expect. Of course this does not
matter if we only want to compute the mean value, but it could have
catastrophic effects in other situations. A safer procedure, is to
simply do ‘mean(mean(patient_data))’. I included the both options.

6. I don’t like to use mathematical-like terms that can be
misinterpreted by lectors with formal mathematics background (and for
these MATLAB lessons, it is likely that we would have such a public).
Consider for example the term ‘perfectly smooth’ used to describe the
plot of ‘max(patient_data,[ ],1)’. The piecewise function that produces
this graph could be many things but not smooth, it is not even
differentiable at x=20!!. What we can say is, for example, that the
maximum value rises and falls at a steady rate.